% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_aggregate.R
\name{call_aggregate}
\alias{call_aggregate}
\title{call_aggregate}
\usage{
call_aggregate(spectra_list, mz_tol, folder_path)
}
\arguments{
\item{spectra_list}{A list of spectra objects for all features, each containing the top x\% TIC spectra concatenated.}

\item{mz_tol}{Mass tolerance (in Da) required when grouping fragments across multiple spectra for a given feature.}

\item{folder_path}{Path to the folder where the data reduction dataframe will be stored.}
}
\value{
A list containing:
\itemize{
\item{\code{aggregate_spectra_object}}: Aggregate spectra object.
\item{\code{aggregate_dataframe}}: Dataframe of aggregate spectra for all features.
}
}
\description{
Helper function to run derive_aggregate_spectra.
}
\examples{
folder_path <- "path/to/folder"
sp <- preprocess(folder_path, 5, 0.01)
l <- sp[[1]]
sp1 <- extract_raw_spectra(folder_path, l, 0.05, 0.8)
spectra_list <- sp1[[1]]
call_aggregate(spectra_list, 0.05, folder_path)

}
