% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_ms2_spec_from_stats_file.R
\name{concat_ms2_spec_from_stats_file}
\alias{concat_ms2_spec_from_stats_file}
\title{concat_ms2_spec_from_stats_file}
\usage{
concat_ms2_spec_from_stats_file(j, stats_file, fl)
}
\arguments{
\item{j}{jth feature from the stats file.}

\item{stats_file}{edited stats file with RT_min and RT_max prefixes added to every sample.}

\item{fl}{list of sample names.}
}
\value{
A list with all spectra concatenated for a particular feature
}
\description{
Helper function for concatenate_spectra(). Need not be called separately. Gets auto-invoked once we call concatenate_spectra()
}
\examples{
# Example usage of the function (need not be called separately)
concat_ms2_spec_from_stats_file(1,stats_file,fl)
}
