% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_spectra.R
\name{concatenate_spectra}
\alias{concatenate_spectra}
\title{concatenate_spectra}
\usage{
concatenate_spectra(output_list_from_preprocess_function, spectral_files)
}
\arguments{
\item{output_list_from_preprocess_function}{A list generated from the preprocess function.}
}
\value{
A list containing the MS2 spectra extracted from the mzML files using the predefined tolerances
and a vector of features for which MS2 spectra couldn't be extracted.
}
\description{
This function extracts MS2 spectra from a list produced by the preprocess function.
}
\details{
Reads a list generated from the preprocess function and extracts MS2 spectra within the given mz and RT range.
Also returns a list of features for which MS2 spectra couldn't be extracted.
}
\examples{
# Example usage of the function
l1 = preprocess(folder_path = "D:/test_2/", tol_mz = 5, tol_rt = 0.1)  # Replace with actual data
concatenate_spectra(l1)

}
