% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_aggregate_spectra.R
\name{derive_aggregate_spectra}
\alias{derive_aggregate_spectra}
\title{derive_aggregate_spectra}
\usage{
derive_aggregate_spectra(sps, mz_tol = 0.05)
}
\arguments{
\item{sps}{A spectra object containing top x\% TIC spectra concatenated for a given feature.}

\item{mz_tol}{mass tolerance (in Da) required when grouping fragments across multiple spectra for a given feature}
}
\value{
A list containing the aggregate spectra as a Spectra object and a dataframe of the aggregate spectra with its mean MZ, Intensity and frequency and the number of fragments before and after inter- and intra-spectra grouping
}
\description{
Create aggregate spectra by grouping fragments across multiple spectra for a given feature using a predefined tolerance and calculate the frequencies of fragments
}
\examples{
# Example usage of the function
derive_aggregate_spectra(sps, 0.05)
}
