% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_matching_before_denoising.R
\name{fragment_matching_before_denoising}
\alias{fragment_matching_before_denoising}
\title{fragment_matching_before_denoising}
\usage{
fragment_matching_before_denoising(folder_path, l4, tolerance, ionization_mode)
}
\arguments{
\item{folder_path}{folder containing the input directory (mzml) feature list files where the results of the matching will be stored.}

\item{l4}{Contain the final set of features which matched with the reference at the MS1 level}

\item{tolerance}{fragment ion tolerance at the MS/MS level, defaults to 0.05 Da}

\item{ionization_mode}{positive or negative modes}
}
\value{
A dataframe with matching metrics, annotations and both experimental and reference spectrum identifiers. Features with matching score zero remain unannotated
}
\description{
Matches experimental and reference spectra using a predefined MS/MS fragment tolerance
}
\examples{
# Example usage of the function
fragment_matching_before_denoising(folder_path, l4, tolerance = 0.05, "positive")
}
