% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrascan_grouping.R
\name{intrascan_grouping}
\alias{intrascan_grouping}
\title{Example function using the pipe operator}
\usage{
intrascan_grouping(scan, tol)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame with the first 10 rows.
}
\description{
Example function using the pipe operator
}
