% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_individual_spectrum.R
\name{label_individual_spectrum}
\alias{label_individual_spectrum}
\title{label_individual_spectrum}
\usage{
label_individual_spectrum(aggregate_list, folder_path, mz_tol)
}
\arguments{
\item{aggregate_list:}{list of aggregate spectra objects for all features (output from the call_aggregate())}

\item{folder_path:}{same path as the one used in preprocess()}

\item{mz_tol:}{mass tolerance (in Da) required when grouping fragments across multiple spectra for a given feature}
}
\value{
a list containing all spectra for a given feature and the corresponding fragment frequencies
}
\description{
Using the frequencies determined from the call_aggregate(), label individual spectrum
}
\examples{
# Example usage of the function
label_individual_spectrum(l2, folder_path, 0.05)
}
