% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precursor_matching.R
\name{precursor_matching}
\alias{precursor_matching}
\title{precursor_matching}
\usage{
precursor_matching(l1, folder_path, ionization_mode, tol = 0.01)
}
\arguments{
\item{l1}{output from the preprocess()}

\item{folder_path}{folder containing the required input directory (mzml) and feature list file in .txt format.}

\item{ionization_mode}{positive or negative. Based on this input, the library data is loaded and stored as a global variable}

\item{tol}{tolerance to group reference spectra fragments, defaults to 0.01}
}
\value{
A dataframe with only those precursors, their mz, RT, etc, which had a corresponding match with the reference library
}
\description{
Matches precursor m/z with library within a given tolerance (As outlined in tol_mz param in preprocess()) and generates a library of reference spectra that can be used for further matching. Also groups fragments in the reference spectra that are close-by within a given tolerance tol
}
\examples{
# Example usage of the function
precursor_matching(l1, folder_path, "positive", 0.05)
}
