% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_reference.R
\name{prepare_reference}
\alias{prepare_reference}
\title{prepare_reference}
\usage{
prepare_reference(index, name, lib, path)
}
\arguments{
\item{index}{index}

\item{name}{feature ID}

\item{lib}{positive or negative library}

\item{path}{folder_path to store the spectra as txt files}
}
\value{
reference spectra in Spectra format
}
\description{
Stores the reference spectra as a txt file for further matching with current experimental spectra and returns the reference spectra as a Spectra object. This neednot be run independently. This is a companion function to precursor_matching.R
}
\examples{
# Example usage of the function
prepare_reference(ind, n, lib, p)
}
