% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{read_files}
\usage{
read_files(folder_path, tol_mz, tol_rt)
}
\arguments{
\item{folder_path:}{same folder path used in preprocess(). The extracted fragment-grouped spectra in .txt format will be stored here}

\item{tol_mz:}{mass tolerance (default: 0.05 Da)}

\item{tol_rt:}{Rt tolerance (default: 0.1667 minutes)}
}
\value{
a list containing all MS2 spectra for every sample, an edited feature list file for subsequent steps and names of samples analyzed
}
\description{
Helper function for preprocess(). Need not be run separately. Runs several checks to see if features file has required columns or if user-defined features file has been provided
Two options are available for inputting feature list from the user. First, the file titled feature_list.txt must be present in the folder_path with the columns "ID", "mz" and "RT". Second, if an additional file titled Sample-and-feature-wise-RT-tolerance.txt is present, it should contain the same number of rows and double the number of columns (RT_min_ and RT_max_ for every sample) as feature_list.txt.
}
\examples{
# Example usage of the function
read_files(folder_path, 0.05, 0.1667)
}
