% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_preparation_for_matching.R
\name{reference_preparation_for_matching}
\alias{reference_preparation_for_matching}
\title{reference_preparation_for_matching}
\usage{
reference_preparation_for_matching(j, annotation_file, lib, path)
}
\arguments{
\item{j}{index}

\item{annotation_file}{stats file with ID, mz, RT etc}

\item{lib}{positive or negative library}

\item{path}{folder_path to store the spectra as txt files}
}
\value{
reference spectra in Spectra format
}
\description{
Prepares the reference library for matching with current experimental spectra. This neednot be run independently. This is a companion function to precursor_matching.R
}
\examples{
# Example usage of the function
reference_preparation_for_matching(j, l1$stats_file_ms2_only, lib, p)
}
