% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{sensitivity_analysis}
\alias{sensitivity_analysis}
\title{sensitivity analysis}
\usage{
sensitivity_analysis(l7, folder_path)
}
\arguments{
\item{l7}{Contain the final set of features with matching metrics before and after denoising}

\item{folder_path}{where the user wants to store the sensitivity plots}
}
\value{
A dataframe with significance values for pairwise comparisons in matching metrics across different thresholds
}
\description{
sensitivity analysis
}
\examples{
# Example usage of the function
sensitivity_analysis(folder_path, l7)
}
