% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_parameters.R
\name{tune_parameters}
\alias{tune_parameters}
\title{tune_parameters}
\usage{
tune_parameters(agg_df, threshold, name, path)
}
\arguments{
\item{agg_df:}{spectrum dataframe with columns Fragment_ID, MZ, Intensity & Freq}

\item{threshold:}{see function description for generate_denoised_spectra}

\item{name:}{names of the spectra such as "scan_1.txt" "scan_2.txt" and so on}

\item{path:}{folder path to store the denoised spectra in .txt and .mzML format}
}
\value{
directly prints the spectra to the file in .txt format and returns the spectrum-wise mz and intensity
}
\description{
Helper function for generate_denoise_spectra(). Need not be called separately
}
\examples{
# Example usage of the function
tune_parameters(agg_df, threshold, "scan_1.txt", path)
}
