% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_aggregate.R
\name{call_aggregate}
\alias{call_aggregate}
\title{derive_aggregate_spectra}
\usage{
call_aggregate(spectra_list, mz_tol, folder_path)
}
\arguments{
\item{spectra_list:}{list of spectra objects for all features, each containing top x\% TIC spectra concatenated}

\item{mz_tol:}{mass tolerance (in Da) required when grouping fragments across multiple spectra for a given feature}

\item{folder_path:}{the path to the folder where you want to store the data reduction dataframe}
}
\value{
a list containing the aggregate spectra object and a dataframe of the aggregate spectra for all features
}
\description{
Helper code to run derive_aggregate_spectra
}
\examples{
# Example usage of the function
call_aggregate(sps_top_tic, 0.05)
}
