% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_spectra.R
\name{concatenate_spectra}
\alias{concatenate_spectra}
\title{concatenate_spectra}
\usage{
concatenate_spectra(output_list_from_preprocess_function)
}
\arguments{
\item{output_list_from_preprocess_function:}{contain list generated from the preprocess()}
}
\value{
A list containing the MS2 spectra extracted from the mzml files using the predefined tolerances and a vector of features for which MS2 spectra couldn't be extracted
}
\description{
Reads a list generated from the preprocess() and extracts MS2 spectra within the given mz and RT range. Also returns a list of features for which MS2 spectra couldn't be extracted
}
\examples{
# Example usage of the function
concatenate_spectra(output_list_from_preprocess_function)
}
