% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_aggregate_spectra.R
\name{derive_aggregate_spectra}
\alias{derive_aggregate_spectra}
\title{derive_aggregate_spectra}
\usage{
derive_aggregate_spectra(sps, mz_tol = 0.05)
}
\arguments{
\item{sps:}{named list of spectra objects containing MS2 data}

\item{mz_tol:}{mass tolerance (in Da) required when grouping fragments across multiple spectra for a given feature}
}
\value{
a list containing the aggregate spectra as a Spectra object and a dataframe of the aggregate spectra with its mean MZ, Intensity and frequency
}
\description{
Create aggregate spectra by grouping fragments across multiple spectra for a given feature using a predefined tolerance and calculate the frequencies of fragments
}
\examples{
# Example usage of the function
derive_aggregate_spectra(sps, 0.05)
}
