% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_denoised_spectra.R
\name{generate_denoised_spectra}
\alias{generate_denoised_spectra}
\title{generate_denoised_spectrum}
\usage{
generate_denoised_spectra(aggregate_list, folder_path, custom_threshold = NULL)
}
\arguments{
\item{aggregate_list:}{output from label_individual_spectrum()}

\item{folder_path:}{same path as the one used in preprocess()}

\item{custom_threshold:}{must be between 0 and 1 (default = NULL). If the user wants to apply a threshold different from ours (0.12 (95\%CI- 0.08-0.15)) they can do so using this parameter. The threshold when multiplied by the number of replicate spectra (for a given feature) will tell us that only fragments with frequencies above this value will be considered as signal fragments and retained in the spectrum}
}
\value{
directly prints the spectra to the file in .txt and .mzML format
}
\description{
Takes in a list of individual spectrum with fragments labeled with frequencies, denoises it using fixed cutoff and writes the resulting spectrum into a file
}
\examples{
# Example usage of the function
generate_denoised_spectra(l3, folder_path, NULL)
}
