% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrascan_grouping.R
\name{intrascan_grouping}
\alias{intrascan_grouping}
\title{intrascan_grouping}
\usage{
intrascan_grouping(scan, tol)
}
\arguments{
\item{tol}{A tolerance value in daltons}

\item{data}{A data frame of a spectrum containing two columns (fragments and intensity).}
}
\value{
A data frame with the same columns but mass fragments grouped.
}
\description{
A function that accepts a spectrum as input and returns a spectrum with fragments grouped within a given mass tolerance
}
