% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{read_files}
\usage{
read_files(folder_path, tol_mz, tol_rt)
}
\arguments{
\item{folder_path:}{same folder path used in preprocess(). The extracted fragment-grouped spectra in .txt format will be stored here}

\item{tol_mz:}{mass tolerance (default: 0.05 Da)}

\item{tol_rt;}{Rt tolerance (default: 0.1667 minutes)}
}
\value{
a list containing all MS2 spectra for every sample, an edited stats file for subsequent steps and names of samples analyzed
}
\description{
Helper function for preprocess(). Need not be run separately. Runs several checks to see if Stats file has required columns or if user-defined Stats file has been provided
}
\examples{
# Example usage of the function
read_files(folder_path, 0.05, 0.1667)
}
